# Preserve non-decodable %-sequences intact when unquote.