--- a/CHANGES/517.bugfix
+++ b/CHANGES/517.bugfix
@@ -1,0 +1,1 @@
+No longer loss characters when decode incorrect percent-sequences (like ``%e2%82%f8``). All non-decodable percent-sequences are now preserved intact.