--- a/requirements/lint.txt
+++ b/requirements/lint.txt
@@ -1,5 +1,5 @@
 mypy==0.790; implementation_name=="cpython"
 flake8==3.8.4
 flake8-pyi==20.5.0; python_version >= "3.6"
-black==20.8b1; python_version >= "3.6"; implementation_name=="cpython"
+black==20.8b1; python_version >= "3.6" and implementation_name=="cpython"
 mypy-extensions==0.4.3; implementation_name=="cpython"