--- a/setup.cfg
+++ b/setup.cfg
@@ -12,7 +12,7 @@ junit_suite_name = yarl_test_suite
 
 
 [flake8]
-ignore = E301,E302,E704,W503,W504,F811
+ignore = E203,E301,E302,E704,W503,W504,F811
 max-line-length = 88
 
 [mypy]