--- a/yarl/_quoting_py.py
+++ b/yarl/_quoting_py.py
@@ -1,3 +1,4 @@
+import codecs
 import re
 from string import ascii_letters, ascii_lowercase, digits
 from typing import Optional, cast
@@ -15,6 +16,8 @@ ALLOWED = UNRESERVED + SUB_DELIMS_WITHOUT_QS
 
 _IS_HEX = re.compile(b"[A-Z0-9][A-Z0-9]")
 _IS_HEX_STR = re.compile("[A-Fa-f0-9][A-Fa-f0-9]")
+
+utf8_decoder = codecs.getincrementaldecoder("utf-8")
 
 
 class _Quoter:
@@ -127,19 +130,30 @@ class _Unquoter:
             raise TypeError("Argument should be str")
         if not val:
             return ""
-        last_pct = ""
-        pcts = bytearray()
+        decoder = cast(codecs.BufferedIncrementalDecoder, utf8_decoder())
         ret = []
         idx = 0
         while idx < len(val):
             ch = val[idx]
             idx += 1
-            if pcts:
-                try:
-                    unquoted = pcts.decode("utf8")
-                except UnicodeDecodeError:
-                    pass
-                else:
+            if ch == "%" and idx <= len(val) - 2:
+                pct = val[idx : idx + 2]
+                if _IS_HEX_STR.fullmatch(pct):
+                    b = bytes([int(pct, base=16)])
+                    idx += 2
+                    try:
+                        unquoted = decoder.decode(b)
+                    except UnicodeDecodeError:
+                        start_pct = idx - 3 - len(decoder.buffer) * 3
+                        ret.append(val[start_pct : idx - 3])
+                        decoder.reset()
+                        try:
+                            unquoted = decoder.decode(b)
+                        except UnicodeDecodeError:
+                            ret.append(val[idx - 3 : idx])
+                            continue
+                    if not unquoted:
+                        continue
                     if self._qs and unquoted in "+=&;":
                         to_add = self._qs_quoter(unquoted)
                         if to_add is None:  # pragma: no cover
@@ -152,19 +166,12 @@ class _Unquoter:
                         ret.append(to_add)
                     else:
                         ret.append(unquoted)
-                    del pcts[:]
-
-            if ch == "%" and idx <= len(val) - 2:
-                pct = val[idx : idx + 2]  # noqa: E203
-                if _IS_HEX_STR.fullmatch(pct):
-                    pcts.append(int(pct, base=16))
-                    last_pct = "%" + pct
-                    idx += 2
                     continue
 
-            if pcts:
-                ret.append(last_pct)  # %F8ab
-                last_pct = ""
+            if decoder.buffer:
+                start_pct = idx - 1 - len(decoder.buffer) * 3
+                ret.append(val[start_pct : idx - 1])
+                decoder.reset()
 
             if ch == "+":
                 if not self._qs or ch in self._unsafe:
@@ -182,24 +189,9 @@ class _Unquoter:
 
             ret.append(ch)
 
-        if pcts:
-            try:
-                unquoted = pcts.decode("utf8")
-            except UnicodeDecodeError:
-                ret.append(last_pct)  # %F8
-            else:
-                if self._qs and unquoted in "+=&;":
-                    to_add = self._qs_quoter(unquoted)
-                    if to_add is None:  # pragma: no cover
-                        raise RuntimeError("Cannot quote None")
-                    ret.append(to_add)
-                elif unquoted in self._unsafe:
-                    to_add = self._qs_quoter(unquoted)
-                    if to_add is None:  # pragma: no cover
-                        raise RuntimeError("Cannot quote None")
-                    ret.append(to_add)
-                else:
-                    ret.append(unquoted)
+        if decoder.buffer:
+            ret.append(val[-len(decoder.buffer) * 3 :])
+
         ret2 = "".join(ret)
         if ret2 == val:
             return val