Closes #517.