@serhiy-storchaka 2020-10-13
The downside is the loss of performance.
Before:
```
Cython unquote: 0.305 sec
Python unquote: 2.103 sec
```
After:
```
Cython unquote: 0.753 sec
Python unquote: 2.700 sec
```