@serhiy-storchaka 2020-10-13
Fixed the performance regression in Cython:
```
Cython unquote: 0.266 sec
Python unquote: 2.615 sec
```